feat: tokenize midi
